perf: build timing issues
