+++ apps/docs/features/docs/Reference.utils.ts
@@ -3,16 +3,167 @@ import { fromMarkdown } from 'mdast-util-from-markdown'
 import { toMarkdown } from 'mdast-util-to-markdown'
 import { mdxFromMarkdown, mdxToMarkdown } from 'mdast-util-mdx'
 import { mdxjs } from 'micromark-extension-mdxjs'
+import type { Metadata, ResolvingMetadata } from 'next'
+import { redirect } from 'next/navigation'
 import { readFile } from 'node:fs/promises'
 import { basename, extname, join } from 'node:path'
 import { visit } from 'unist-util-visit'
 import { parse } from 'yaml'
 
+import { REFERENCES, clientSdkIds } from '~/content/navigation.references'
 import { deepFilterRec } from '~/features/helpers.fn'
 import type { Json } from '~/features/helpers.types'
 import { cache_fullProcess_withDevCacheBust } from '~/features/helpers.fs'
+import { generateOpenGraphImageMeta } from '~/features/seo/openGraph'
+import { BASE_PATH } from '~/lib/constants'
 import { SPEC_DIRECTORY } from '~/lib/docs'
 import commonClientLibSections from '~/spec/common-client-libs-sections.json' assert { type: 'json' }
+
+export function parseReferencePath(slug: Array<string>) {
+  const isClientSdkReference = clientSdkIds.includes(slug[0])
+
+  if (isClientSdkReference) {
+    let [sdkId, maybeVersion, maybeCrawlers, ...path] = slug
+    if (!/v\d+/.test(maybeVersion)) {
+      maybeVersion = null
+      maybeCrawlers = maybeVersion
+      path = [maybeCrawlers, ...path]
+    }
+    if (maybeCrawlers !== 'crawlers') {
+      maybeCrawlers = null
+      path = [maybeCrawlers, ...path]
+    }
+
+    return {
+      __type: 'clientSdk' as const,
+      sdkId,
+      maybeVersion,
+      maybeCrawlers,
+      path,
+    }
+  } else {
+    return {
+      __type: 'UNIMPLEMENTED' as const,
+    }
+  }
+}
+
+async function generateStaticParamsForSdkVersion(
+  sdkId: string,
+  version: string,
+  maybeSpecFile?: string,
+  maybeExcludeName?: string
+) {
+  const specFile = maybeSpecFile ?? REFERENCES[sdkId].meta[version].specFile
+  const excludeName = maybeExcludeName ?? REFERENCES[sdkId].meta[version].excludeName
+
+  const sectionTree = await genClientSdkSectionTree(specFile, excludeName)
+  const flattenedSections = flattenCommonClientLibSections(sectionTree)
+
+  return flattenedSections
+    .filter((section) => section.type !== 'category' && !!section.slug)
+    .map((section) => ({
+      slug: [
+        sdkId,
+        version === REFERENCES[sdkId].versions[0] ? null : version,
+        'crawlers',
+        section.slug,
+      ].filter(Boolean),
+    }))
+}
+
+export async function generateReferenceStaticParams() {
+  const crawlerPages = clientSdkIds
+    .flatMap((sdkId) =>
+      REFERENCES[sdkId].versions.map((version) => ({
+        sdkId,
+        version,
+        specFile: REFERENCES[sdkId].meta[version].specFile,
+        excludeName: REFERENCES[sdkId].meta[version].excludeName,
+      }))
+    )
+    .map(async ({ sdkId, version, specFile, excludeName }) => {
+      return generateStaticParamsForSdkVersion(sdkId, version, specFile, excludeName)
+    })
+
+  const nonCrawlerPages = clientSdkIds
+    .flatMap((sdkId) =>
+      REFERENCES[sdkId].versions.map((version) => ({
+        sdkId,
+        version,
+      }))
+    )
+    .map(({ sdkId, version }) => ({
+      slug: [sdkId, version === REFERENCES[sdkId].versions[0] ? null : version].filter(Boolean),
+    }))
+
+  return nonCrawlerPages.concat((await Promise.all(crawlerPages)).flat())
+}
+
+export async function generateReferenceMetadata(
+  { params: { slug } }: { params: { slug: Array<string> } },
+  resolvingParent: ResolvingMetadata
+): Promise<Metadata> {
+  const { alternates: parentAlternates, openGraph: parentOg } = await resolvingParent
+  const flattenedSections = flattenCommonClientLibSections(
+    commonClientLibSections as Array<AbbrevCommonClientLibSection>
+  )
+
+  const parsedPath = parseReferencePath(slug)
+  const isClientSdkReference = parsedPath.__type === 'clientSdk'
+
+  if (isClientSdkReference) {
+    const { sdkId, maybeVersion } = parsedPath
+
+    const displayName = REFERENCES[sdkId].name
+    const sectionTitle =
+      slug.length > 0
+        ? flattenedSections.find((section) => section.slug === slug[0])?.title
+        : undefined
+    const url = [BASE_PATH, 'reference', sdkId, maybeVersion, slug[0]].filter(Boolean).join('/')
+
+    return {
+      title: `${displayName} API Reference | Supabase Docs`,
+      description: `API reference for the ${displayName} Supabase SDK`,
+      ...(slug.length > 0
+        ? {
+            alternates: {
+              ...parentAlternates,
+              canonical: url,
+            },
+          }
+        : {}),
+      openGraph: {
+        ...parentOg,
+        url,
+        images: generateOpenGraphImageMeta({
+          type: 'API Reference',
+          title: `${displayName}${sectionTitle ? `: ${sectionTitle}` : ''}`,
+        }),
+      },
+    }
+  } else {
+    return {}
+  }
+}
+
+export async function redirectNonexistentReferenceSection(
+  sdkId: string,
+  version: string,
+  path: Array<string>,
+  isLatestVersion: boolean
+) {
+  const initialSelectedSection = path[0]
+
+  const validSlugs = await generateStaticParamsForSdkVersion(sdkId, version)
+
+  if (
+    initialSelectedSection &&
+    !validSlugs.some((params) => params.slug[0] === initialSelectedSection)
+  ) {
+    redirect(`/reference/${sdkId}` + (!isLatestVersion ? '/' + version : ''))
+  }
+}
 
 interface AbbrevCommonClientLibSection {
   id: string
