+++ apps/docs/middleware.ts
@@ -5,7 +5,7 @@ import { NextResponse } from 'next/server'
 import { clientSdkIds } from '~/content/navigation.references'
 import { BASE_PATH } from '~/lib/constants'
 
-const REFERENCE_PATH = `${BASE_PATH ?? ''} + '/'}reference`
+const REFERENCE_PATH = `${(BASE_PATH ?? '') + '/'}reference`
 
 export function middleware(request: NextRequest) {
   const url = new URL(request.url)
@@ -32,9 +32,10 @@ export function middleware(request: NextRequest) {
       }
     }
   } else {
-    const [, lib, ...slug] = url.pathname.replace(REFERENCE_PATH, '').split('/')
+    const [, lib, maybeVersion] = url.pathname.replace(REFERENCE_PATH, '').split('/')
     if (clientSdkIds.includes(lib)) {
-      const rewritePath = [REFERENCE_PATH, lib, ...slug].join('/')
+      const version = /v\d+/.test(maybeVersion) ? maybeVersion : null
+      const rewritePath = [REFERENCE_PATH, lib, version].filter(Boolean).join('/')
       return NextResponse.rewrite(new URL(rewritePath, request.url))
     }
   }
